New MoveBrain NinjaMove
NinjaMove looks to move around other robots to points where fewer robots
can shoot at it. The goal is to have it line up two other robots so that
only one of the robots can shoot at directly. If it does this with
multiple pairs of robots, it reduces the chances of it getting hit, in
theory. To be tested