Fix to the ninja move code
Still a bit of a problem, but now for the most part it moves where it is
told now. Prior to this, it had a unit confusion problem (between
degrees and radians), so it would run away up into a corner. Now for the
most part it hides behind another robot, although it lags by up to 8
ticks, plus movement time, which should be fixed by improving the radar.
Next goal is to fix the circling problem when it reaches where it wants
to go.