New architecture for future robots
The robot is cwruBase. It has various Brains for each robot function
(movement, radar, gun) and a "box" to store all of the data that is
collected.
The box stores data by the original brain that created the data. All of
the data can be accessed by requesting a specific brain's data or by
requesting a generic brain type (ex. Sonar). From there, the data is
sorted by Array with an attached string identifier.
The basic brains (names unofficial at this point) merely relay data to
be stored and data as it is accessed. Extending those brains will allow
the robot to take actions.