Sonar on the new system
Sonar is up to date with minor changes to get it talking in RoboCore and
RobotBits. The three brains Gunner, Legs and Sonar now all have [final]
set methods. To modify the brains, extend the G/L/S brain, and modify
the process method. This way there will always be a value to set to
regardless if the brains time out.